feature enhance current weather ui design

+++ app/src/main/java/com/example/weatherapp/featureweather/data/remote/dto/CountryDto.kt
@@ -2,6 +2,6 @@ package com.example.weatherapp.featureweather.data.remote.dto
 
 data class CountryDto(
     val country: String,
-    val sunrise: Float,
-    val sunset: Float,
+    val sunrise: Long,
+    val sunset: Long,
 )
